Make data depends on domain

+++ Timer/src/main/java/com/pomodorotime/timer/models/TimerEvents.kt
@@ -1,7 +1,7 @@
 package com.pomodorotime.timer.models
 
 sealed class TimerEvents {
-    data class LoadData(val id: Int) : TimerEvents()
+    data class LoadData(val id: Long) : TimerEvents()
     object OnPlayStopButtonClicked: TimerEvents()
     object OnBackPressed: TimerEvents()
 }
